Changes webChooser pattern
